pipeline: move AstroPix impl to a separate module
