--- docs/api/toasty.pipeline.AstroPixCandidateInput.rst
@@ -1,19 +1,0 @@
-AstroPixCandidateInput
-======================
-
-.. currentmodule:: toasty.pipeline
-
-.. autoclass:: AstroPixCandidateInput
-   :show-inheritance:
-
-   .. rubric:: Methods Summary
-
-   .. autosummary::
-
-      ~AstroPixCandidateInput.cache_data
-      ~AstroPixCandidateInput.get_unique_id
-
-   .. rubric:: Methods Documentation
-
-   .. automethod:: cache_data
-   .. automethod:: get_unique_id
