--- docs/api/toasty.pipeline.AstroPixImageSource.rst
@@ -1,23 +1,0 @@
-AstroPixImageSource
-===================
-
-.. currentmodule:: toasty.pipeline
-
-.. autoclass:: AstroPixImageSource
-   :show-inheritance:
-
-   .. rubric:: Methods Summary
-
-   .. autosummary::
-
-      ~AstroPixImageSource.deserialize
-      ~AstroPixImageSource.get_config_key
-      ~AstroPixImageSource.open_input
-      ~AstroPixImageSource.query_candidates
-
-   .. rubric:: Methods Documentation
-
-   .. automethod:: deserialize
-   .. automethod:: get_config_key
-   .. automethod:: open_input
-   .. automethod:: query_candidates
