--- docs/api/toasty.pipeline.AstroPixInputImage.rst
@@ -1,43 +1,0 @@
-AstroPixInputImage
-==================
-
-.. currentmodule:: toasty.pipeline
-
-.. autoclass:: AstroPixInputImage
-   :show-inheritance:
-
-   .. rubric:: Attributes Summary
-
-   .. autosummary::
-
-      ~AstroPixInputImage.global_id
-      ~AstroPixInputImage.image_id
-      ~AstroPixInputImage.lower_id
-      ~AstroPixInputImage.normalized_extension
-      ~AstroPixInputImage.publisher_id
-      ~AstroPixInputImage.resource_url
-      ~AstroPixInputImage.wcs_coordinate_frame
-      ~AstroPixInputImage.wcs_equinox
-      ~AstroPixInputImage.wcs_projection
-      ~AstroPixInputImage.wcs_reference_dimension
-      ~AstroPixInputImage.wcs_reference_pixel
-      ~AstroPixInputImage.wcs_reference_value
-      ~AstroPixInputImage.wcs_rotation
-      ~AstroPixInputImage.wcs_scale
-
-   .. rubric:: Attributes Documentation
-
-   .. autoattribute:: global_id
-   .. autoattribute:: image_id
-   .. autoattribute:: lower_id
-   .. autoattribute:: normalized_extension
-   .. autoattribute:: publisher_id
-   .. autoattribute:: resource_url
-   .. autoattribute:: wcs_coordinate_frame
-   .. autoattribute:: wcs_equinox
-   .. autoattribute:: wcs_projection
-   .. autoattribute:: wcs_reference_dimension
-   .. autoattribute:: wcs_reference_pixel
-   .. autoattribute:: wcs_reference_value
-   .. autoattribute:: wcs_rotation
-   .. autoattribute:: wcs_scale
